⚗ Moved all passport logic
